+++ docs/changelog.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Logo
+
+### Fixed
+
+* Accessor `is_valid` did not return `False` when a dataset does not comply 
+  with the schema.
 ### Removed
 
 * Obsolete test data